feat: add confirmation before deleting a location

+++ js/app.controller.js
@@ -69,16 +69,14 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    if (!confirm('Are you sure you want to delete this location?')) return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
             unDisplayLoc()
             loadAndRenderLocs()
         })
-        .catch(err => {
-            console.error('OOPs:', err)
-            flashMsg('Cannot remove location')
-        })
+        .catch(() => flashMsg('Cannot remove location'))
 }
 
 function onSearchAddress(ev) {
@@ -223,7 +221,7 @@ function getFilterByFromQueryParams() {
     const queryParams = new URLSearchParams(window.location.search)
     const txt = queryParams.get('txt') || ''
     const minRate = queryParams.get('minRate') || 0
-    locService.setFilterBy({txt, minRate})
+    locService.setFilterBy({ txt, minRate })
 
     document.querySelector('input[name="filter-by-txt"]').value = txt
     document.querySelector('input[name="filter-by-rate"]').value = minRate
